feat(reth): Added botanix-configs and extended botanix-cli-args (#881)
