+++ Cross.toml
@@ -1,12 +1,12 @@
 [build]
 pre-build = [
-  "curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y",
-  "apt-get update && apt-get install --assume-yes --no-install-recommends build-essential llvm-dev protobuf-compiler gcc pkg-config libprotobuf-dev libclang-6.0-dev clang-6.0 curl", # install system deps                                                                    # add target
-  "ln -s /root/.cargo/bin/cargo /usr/local/bin/cargo",
-  "ln -s /root/.cargo/bin/rustc /usr/local/bin/rustc",
-  "ln -s /root/.cargo/bin/rustup /usr/local/bin/rustup",
-  "/root/.cargo/bin/rustup target add aarch64-unknown-linux-gnu",
-  "/root/.cargo/bin/rustup target add x86_64-unknown-linux-gnu",
+    "curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y",
+    "apt-get update && apt-get install --assume-yes --no-install-recommends build-essential llvm-dev protobuf-compiler gcc pkg-config libprotobuf-dev libclang-6.0-dev clang-6.0 curl", # install system deps                                                                    # add target
+    "ln -s /root/.cargo/bin/cargo /usr/local/bin/cargo",
+    "ln -s /root/.cargo/bin/rustc /usr/local/bin/rustc",
+    "ln -s /root/.cargo/bin/rustup /usr/local/bin/rustup",
+    "/root/.cargo/bin/rustup target add aarch64-unknown-linux-gnu",
+    "/root/.cargo/bin/rustup target add x86_64-unknown-linux-gnu",
 ]
 
 [build.env]
