+++ bin/botanix-up/Cargo.toml
@@ -15,6 +15,7 @@ path = "src/main.rs"
 [dependencies]
 anyhow = "1.0.95"
 botanix-cli-args = { workspace = true }
+botanix-configs = { workspace = true }
 clap = { version = "4.5", features = ["derive"] }
 hex = "0.4.3"
 rand = { workspace = true }
@@ -28,5 +29,6 @@ tokio = { version = "1.45.0", features = ["full"] }
 toml = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
+
 [lints]
 workspace = true
