+++ bin/botanix-up/src/main.rs
@@ -4,7 +4,7 @@
 mod cli;
 use crate::comet_node::{get_enode, TestSignal};
 use anyhow::{Context, Result as AnyResult};
-use botanix_cli_args::federation_args::{FedMemberPubKey, FederationTomlConfig};
+use botanix_configs::federation::{FedMemberPubKey, FederationTomlConfig};
 use clap::Parser;
 use cli::Cli;
 use reth_node_core::primitives::Address;
