+++ bin/test-suite/Cargo.toml
@@ -30,6 +30,7 @@ botanix-authority-peg = { workspace = true, features = ["secp256k1"] }
 botanix-btc-wallet = { path = "../../crates/botanix-btc-wallet" }
 botanix-cli-args = { workspace = true }
 botanix-comet-bft-rpc = { workspace = true }
+botanix-configs = { workspace = true }
 botanix-data-parser = { workspace = true }
 botanix-storage = { workspace = true }
 btc-server = { path = "../btc-server" }
