+++ crates/botanix-activation-manager/Cargo.toml
@@ -7,12 +7,11 @@ license = { workspace = true }
 
 [dependencies]
 botanix-storage = { workspace = true }
-reth-storage-errors = { workspace = true }
 reth-primitives = { workspace = true }
+reth-storage-errors = { workspace = true }
 
 [dev-dependencies]
 botanix-activation-manager = { workspace = true, features = ["test-utils"] }
-
 
 [features]
 test-utils = []
