+++ crates/botanix-cli-args/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod bitcoind;
-pub mod federation_args;
+pub mod chain;
 pub mod frost_args;
 pub mod poa_node;
 pub mod state_sync;
