+++ crates/consensus/authority/Cargo.toml
@@ -15,6 +15,7 @@ async-trait = { workspace = true }
 
 ## Bitcoin
 bitcoin = { workspace = true }
+botanix-activation-manager = { workspace = true }
 botanix-authority-edh = { workspace = true }
 
 ## Botanix
@@ -27,7 +28,6 @@ botanix-cli-args = { workspace = true }
 botanix-comet-bft-rpc = { workspace = true }
 botanix-data-parser = { workspace = true }
 botanix-storage = { workspace = true }
-botanix-activation-manager = { workspace = true }
 btc-server = { path = "../../../bin/btc-server" }
 
 # others
