+++ crates/node/core/src/args/mod.rs
@@ -56,6 +56,4 @@ pub use datadir_args::DatadirArgs;
 mod benchmark_args;
 pub use benchmark_args::BenchmarkArgs;
 
-pub mod utils;
-
 pub mod types;
